requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
praw==7.7.1
prawcore==2.4.0
python-dotenv==1.0.0
requests==2.31.0
update-checker==0.18.0
urllib3==2.0.6
websocket-client==1.6.4